# api/openapi.yaml
openapi: 3.1.0
info:
  title: Squyrrl API
  version: 0.1.0
  description: |
    Squyrrl（箩筐）主 API 契约。本文件描述 Phase 1 已实现的 20 个端点。

    ## 认证模型

    - 客户端用邮箱 OTP 登录得到 `access_token`（1h）+ `refresh_token`（30d）。
    - 受保护端点要求 `Authorization: Bearer <access_token>` 请求头。
    - access token 过期后用 `/auth/refresh` 旋转出新的 access；refresh token 不旋转。
    - 任何 token 都仅以 SHA-256 哈希入库；明文 token 永不持久化。

    ## 同步并发控制

    - `snippets.version` 是单调自增的乐观锁。
    - 写入端点（PATCH / DELETE）必须携带 `If-Match: "<version>"` 头，否则 400 / 409。
    - 读取端点会返回 `ETag: "<version>"` 响应头。
    - 增量同步：`GET /snippets?since=<RFC3339>` 返回 `updated_at >= since` 的碎片，按时间升序，附带 `next_since` 用于下一次请求。
    - `?include_deleted=true` 让同步流包含软删除的碎片（带 `deleted_at`），客户端据此本地下架。

    ## 数据混淆

    - `snippets.text_content` 与未来的文件 blob 在客户端用 XChaCha20-Poly1305 + 应用静态密钥混淆后再上传。
    - 服务器只见混淆字节，不解密。在 JSON 中字段以 base64 出现（Go `[]byte` 默认编码）。
  contact:
    name: Squyrrl
    url: https://squyrrl.app
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: 开发环境（本机 Docker Compose）
  - url: https://api.squyrrl.app
    description: 生产环境（占位）
...
security:
  - bearerAuth: []

# =============================================================================
# 端点
# =============================================================================
paths:
...
  /auth/email/request:
    post:
      operationId: requestEmailOtp
      tags: [Auth]
      summary: 请求邮箱登录验证码
      description: |
        生成 6 位数字 OTP，10 分钟有效，通过 SMTP 邮件投递。
        无论邮箱是否存在均返回 204，避免邮箱枚举。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '204': { description: 已（尝试）发送 }
        '400': { $ref: '#/components/responses/BadRequest' }
  /auth/email/verify:
    post:
      operationId: verifyEmailOtp
      tags: [Auth]
      summary: 用 OTP 验证并签发会话
      description: 校验 OTP 成功后注册或登录用户，注册或更新设备，签发 access + refresh token。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code, device]
              properties:
                email: { type: string, format: email }
                code: { type: string, pattern: '^\d{6}$' }
                device: { $ref: '#/components/schemas/DeviceInput' }
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [Auth]
      summary: 旋转 access token
      description: 用 refresh_token 换发新的 access_token。refresh_token 本身不旋转。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: 新 token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      operationId: logout
      tags: [Auth]
      summary: 登出当前会话
      description: 撤销当前 access token 对应的会话；客户端应清空本地 token。
      responses:
        '204': { description: 已撤销 }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [Auth]
      summary: 当前用户信息
      responses:
        '200':
          description: 当前 user + device 视角
          content:
            application/json:
              schema:
                type: object
                required: [user, device_id]
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  device_id: { type: string, format: uuid }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ---------------------------------------------------------------- Pages
...
  /snippets/{id}/archive:
    parameters:
      - $ref: '#/components/parameters/UUIDPath'
    post:
      operationId: archiveSnippet
      tags: [Snippets]
      summary: 抓取 URL 碎片原始内容并作为附件入库
      description: |
        - 仅适用于 `type=url` 的碎片
        - 服务端用 `net/http` 抓取 `payload.url` 一次（5 MiB 上限），**不执行 JS、不内联资源**
          —— ADR-006 完整 SingleFile 留作 Phase 2（chromedp / headless Chrome）
        - 抓到的字节作为 `text/html`（或上游 Content-Type）上传到 files 表
        - 创建一条 element 记录挂在该 snippet 下，alt_text 固定为 `Archived HTML`
        - snippet.version 自增 1，让客户端通过 sync 流拉到新 element
        - 同一 URL 多次 archive：files 去重，但仍会插入新 element（用户可手动删除）
      responses:
        '200':
          description: 已归档
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArchiveResponse' }
        '400':
          description: snippet 非 URL 类型 / payload 缺 url 字段 / 解析 ID 失败
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413':
          description: 目标内容超过 5MiB
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '502':
          description: 抓取上游失败
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ---------------------------------------------------------------- Files
...
  /auth/passkey/register/begin:
    post:
      operationId: passkeyRegisterBegin
      tags: [Passkey]
      summary: 已登录用户：发起 passkey 注册
      description: |
        返回 PublicKeyCredentialCreationOptions（W3C Web Authentication）+ session_id。
        客户端应调用 `navigator.credentials.create({ publicKey })`，
        把生成的 `PublicKeyCredential` 序列化后通过 `register/finish` 回传。
      responses:
        '200':
          description: 注册 ceremony 已开始
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PasskeyRegisterBeginResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/passkey/register/finish:
    post:
      operationId: passkeyRegisterFinish
      tags: [Passkey]
      summary: 已登录用户：提交 passkey 凭据完成注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id, response]
              properties:
                session_id: { type: string }
                response:
                  type: object
                  description: 客户端 PublicKeyCredential 的 JSON 序列化
                  additionalProperties: true
                name:
                  type: string
                  description: 用户给该凭据起的名字（"MacBook 内置 TouchID" 等）
      responses:
        '204': { description: 已绑定 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/passkey/login/begin:
    post:
      operationId: passkeyLoginBegin
      tags: [Passkey]
      summary: 公开：发起 passkey 登录（无密码）
      description: |
        服务端按 email 定位用户并发出 challenge。
        返回 PublicKeyCredentialRequestOptions + session_id；
        客户端调 `navigator.credentials.get({ publicKey })` 后回传到 `login/finish`。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: 登录 ceremony 已开始
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PasskeyLoginBeginResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401':
          description: 用户未注册过 passkey 或邮箱不存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/passkey/login/finish:
    post:
      operationId: passkeyLoginFinish
      tags: [Passkey]
      summary: 公开：提交 passkey assertion，签发会话
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id, email, response, device]
              properties:
                session_id: { type: string }
                email: { type: string, format: email }
                response:
                  type: object
                  additionalProperties: true
                device: { $ref: '#/components/schemas/DeviceInput' }
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ---------------------------------------------------------------- QR 扫码绑定
  /auth/qr/start:
    post:
      operationId: qrStart
      tags: [QR]
      summary: 已登录设备：拿一个 90 秒有效的绑定码
      description: |
        客户端可把返回的 `code` 编码为二维码展示给新设备扫描，
        或直接让用户在新设备手动输入绑定码（推荐 8 字符不易混淆字符集）。
      responses:
        '200':
          description: 绑定码
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QRCode' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/qr/redeem:
    post:
      operationId: qrRedeem
      tags: [QR]
      summary: 新设备：兑换绑定码 + 设备元数据 → 登录
      description: |
        与 `/auth/email/verify` / `/auth/passkey/login/finish` 共享 LoginResult 响应；
        签发到的 access/refresh token 与其它登录方式无差异。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, device]
              properties:
                code:
                  type: string
                  minLength: 8
                  maxLength: 8
                device: { $ref: '#/components/schemas/DeviceInput' }
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401':
          description: 绑定码无效或已过期
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

# =============================================================================
# 复用组件
# =============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: 仅供 Bot 等内部服务调用；与主 API 共享同一个秘密
  parameters:
    UUIDPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: 请求参数无效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: 未认证 / token 无效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: 资源不存在或不归属当前用户
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: 人可读的错误描述（中文） }

    # ---------- 认证 ----------
    User:
      type: object
      required: [id, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        email:
          type: [string, 'null']
          format: email
        handle:
          type: [string, 'null']
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Device:
      type: object
      required: [id, user_id, name, platform, last_seen_at]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        name: { type: string }
        platform: { $ref: '#/components/schemas/Platform' }
        last_seen_at: { type: string, format: date-time }
    Platform:
      type: string
      enum: [ios, android, macos, windows, linux, web, browser, telegram, wechat]
    DeviceInput:
      type: object
      required: [name, platform]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        platform: { $ref: '#/components/schemas/Platform' }
    TokenPair:
      type: object
      required: [access_token, refresh_token, access_expires_at, refresh_expires_at]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        access_expires_at: { type: string, format: date-time }
        refresh_expires_at: { type: string, format: date-time }
    LoginResult:
      type: object
      required: [user, device, tokens]
      properties:
        user: { $ref: '#/components/schemas/User' }
        device: { $ref: '#/components/schemas/Device' }
        tokens: { $ref: '#/components/schemas/TokenPair' }

    # ---------- 页面 ----------
...
    PasskeyRegisterBeginResponse:
      type: object
      required: [publicKey, session_id]
      properties:
        publicKey:
          type: object
          additionalProperties: true
          description: PublicKeyCredentialCreationOptions（W3C 规范，由 webauthn 库直接序列化）
        session_id:
          type: string
          description: 服务器给客户端的临时 ID；finish 时回传以索引 ceremony state
    PasskeyLoginBeginResponse:
      type: object
      required: [publicKey, session_id]
      properties:
        publicKey:
          type: object
          additionalProperties: true
          description: PublicKeyCredentialRequestOptions
        session_id: { type: string }

    # ---------- 冲突归档 ----------
...
    ArchiveResponse:
      type: object
      required: [snippet_id, file_id, element_id, size_bytes, mime]
      properties:
        snippet_id: { type: string, format: uuid }
        file_id: { type: string, format: uuid }
        element_id: { type: string, format: uuid }
        size_bytes:
          type: integer
          format: int64
          description: 抓到的原始字节数（≤ 5 MiB）
        mime: { type: string, examples: [text/html] }

    # ---------- 钱包 ----------
...
    QRCode:
      type: object
      required: [code, expires_at]
      properties:
        code:
          type: string
          minLength: 8
          maxLength: 8
          description: 8 位不易混淆字符（去掉 0/O/1/I/L）
        expires_at: { type: string, format: date-time }

    # ---------- Telegram 绑定 ----------
...